Added files to repository
Added the version 0.9 of CStandardContainers to the repository. Current version supports the containers DynamicArray, LinkedList, AString and WString aswell as the interfaces Allocator and Iterator. Currently only supports Von-Neumann-architectures. Future versions will base on a complete rewrite of the library, which will improve many factors as well as factor out some of the library's restrictions.